Fix Model Type
Cache Addition:

We created the private field private array $enumCache = []; to store the results of the isEnum checks. Before creating a new ReflectionClass object, we check if the result already exists in the cache.

Class Existence Check:

We use the class_exists($cast) function to check if the class we are trying to examine exists. This prevents potential exceptions due to non-existent classes.

Exception Handling:

We added a try-catch block to handle any ReflectionException that may arise during the creation of the ReflectionClass object. In case of an exception, we store the result false in the cache and return false.

Removal of Exception Suppression Comment:

We removed the comment /** @noinspection PhpUnhandledExceptionInspection */ since we now handle the exceptions properly.

Parameter Type:

We set the parameter type $name as string for greater clarity and strictness.