Type Checking in set() Methods:
Type Checking in set() Methods:

In the set() method, when the values entered are either IncomeClassification or ExpensesClassification, there is no check to verify if the item is actually valid before it is added. Invalid values can be inserted without any validation. It would be useful to add checks for the format of the incoming data.